backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.7
django-cors==0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pycodestyle==2.11.1
pytz==2023.3.post1
sqlparse==0.4.4
